# server/controllers/Laboratorio_db/generated/ClientiControllerGenerated.js
/**
 * 
 * 
  _____                      _              _ _ _     _   _     _        __ _ _      
 |  __ \                    | |            | (_) |   | | | |   (_)      / _(_) |     
 | |  | | ___    _ __   ___ | |_    ___  __| |_| |_  | |_| |__  _ ___  | |_ _| | ___ 
 | |  | |/ _ \  | '_ \ / _ \| __|  / _ \/ _` | | __| | __| '_ \| / __| |  _| | |/ _ \
 | |__| | (_) | | | | | (_) | |_  |  __/ (_| | | |_  | |_| | | | \__ \ | | | | |  __/
 |_____/ \___/  |_| |_|\___/ \__|  \___|\__,_|_|\__|  \__|_| |_|_|___/ |_| |_|_|\___|
 
 * DO NOT EDIT THIS FILE!! 
 * 
 *  TO CUSTOMIZE ClientiControllerGenerated.js PLEASE EDIT ../ClientiController.js
 * 
 *  -- THIS FILE WILL BE OVERWRITTEN ON THE NEXT SKAFFOLDER'S CODE GENERATION --
 * 
 */
import Properties from "../../../properties";
import ClientiModel from "../../../models/Laboratorio_db/ClientiModel";
import ErrorManager from "../../../classes/ErrorManager";
import { authorize } from "../../../security/SecurityManager";
import ClientiController from "../ClientiController";

const generatedControllers = {
  /**
   * Init routes
   */
  init: router => {
    const baseUrl = `${Properties.api}/clienti`;
    router.post(baseUrl + "", authorize([]), ClientiController.create);
    router.delete(baseUrl + "/:id", authorize([]), ClientiController.delete);
    router.get(baseUrl + "", authorize([]), ClientiController.list);
    router.post(baseUrl + "/:id", authorize([]), ClientiController.update);
  },


  // CRUD METHODS


  /**
  * ClientiModel.create
  *   @description CRUD ACTION create
  *   @param Clienti obj Object to insert
  *
  */
  create: async (req, res) => {
    try {
      const result = await ClientiModel.create(req.body);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * ClientiModel.delete
  *   @description CRUD ACTION delete
  *   @param ObjectId id Id Clienti
  *
  */
  delete: async (req, res) => {
    try {
      const result = await ClientiModel.delete(req.params.id);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * ClientiModel.list
  *   @description CRUD ACTION list
  *   @returns ARRAY OF Clienti
  *
  */
  list: async (req, res) => {
    try {
      const result = await ClientiModel.list();
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  
  /**
  * ClientiModel.update
  *   @description CRUD ACTION update
  *   @param ObjectId id Id Clienti
  *   @returns Clienti
  *
  */
  update: async (req, res) => {
    try {
      const result = await ClientiModel.update(req.body);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  
  // Custom APIs

};

export default generatedControllers;